+++ apps/contact/app/api/contact/route.ts
@@ -1,5 +1,6 @@
-import { Client, isFullPage } from "@notionhq/client";
+import { processContact } from "@/app/(helpers)/notion";
 import { nanoid } from "nanoid";
+import { NextRequest } from "next/server";
 import z from "zod";
 
 const bodyValidationSchema = z.object({
@@ -7,254 +8,15 @@ const bodyValidationSchema = z.object({
     .string()
     .min(3, { message: "Name must be at least 3 characters long" }),
   email: z
-    .string()
-    .min(1, { message: "Required field" })
-    .email({ message: "Invalid email adress" }),
+    .email({ message: "Invalid email adress" })
+    .min(1, { message: "Required field" }),
   message: z.string().min(1, { message: "Required field" }),
   hasConsent: z.boolean().optional(),
 });
 
 type RequestBody = z.infer<typeof bodyValidationSchema>;
 
-const {
-  NOTION_TOKEN,
-  SLACK_CHANNEL,
-  SLACK_BOT_TOKEN,
-  MENTION_EMAILS,
-  MENTION_IDS,
-  NOTION_DATABASE_ID,
-  IS_OFFLINE,
-} = process.env;
-
-const notion = new Client({ auth: NOTION_TOKEN });
-
-const createPayload = (name: string, email: string, url: string) => ({
-  channel: SLACK_CHANNEL,
-  blocks: [
-    {
-      type: "header",
-      text: {
-        type: "plain_text",
-        text: "We have 1 new message(s).",
-        emoji: true,
-      },
-    },
-    {
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: `We got a new message from _${name}_ (_${email}_).`,
-      },
-    },
-    {
-      type: "divider",
-    },
-    {
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: " ",
-      },
-      accessory: {
-        type: "button",
-        text: {
-          type: "plain_text",
-          text: "Show me the message",
-          emoji: true,
-        },
-        value: "new_message_click",
-        url,
-        action_id: "button-action",
-      },
-    },
-  ],
-});
-
-const notifyContactCreated = async (
-  name: string,
-  email: string,
-  url: string,
-) => {
-  const payload = createPayload(name, email, url);
-  const payloadStringify = JSON.stringify(payload);
-
-  if (IS_OFFLINE) {
-    console.log(payload);
-  } else {
-    try {
-      const result = await fetch("https://slack.com/api/chat.postMessage", {
-        method: "POST",
-        body: payloadStringify,
-        headers: {
-          "Content-Type": "application/json; charset=utf-8",
-          "Content-Length": payloadStringify.length.toString(),
-          Authorization: `Bearer ${SLACK_BOT_TOKEN}`,
-          Accept: "application/json",
-        },
-      });
-      if (result.status !== 200) {
-        throw {
-          statusCode: result.status,
-          headers: {
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Credentials": true,
-          },
-        };
-      }
-    } catch (error) {
-      throw error;
-    }
-  }
-};
-
-const mentionPerson = ({ id, email }: { id: string; email: string }) => [
-  {
-    mention: {
-      user: {
-        id,
-        person: {
-          email,
-        },
-      },
-    },
-    plain_text: "",
-    href: null,
-  },
-  {
-    text: {
-      content: " ",
-    },
-  },
-];
-
-const getMentions = () => {
-  if (MENTION_EMAILS && MENTION_IDS) {
-    const emails = MENTION_EMAILS.split(",");
-    const ids = MENTION_IDS.split(",");
-
-    if (emails.length && ids.length) {
-      return ids.map((id, i) => ({
-        id,
-        email: emails[i],
-      }));
-    }
-  }
-  return [];
-};
-
-const mentionPeople = () => {
-  return getMentions().flatMap(mentionPerson);
-};
-
-const createContactObject = (
-  id: string,
-  email: string,
-  name: string,
-  content: string,
-) => ({
-  parent: {
-    database_id: NOTION_DATABASE_ID || "",
-  },
-  properties: {
-    id: {
-      title: [
-        {
-          text: {
-            content: id,
-          },
-        },
-      ],
-    },
-    email: {
-      email,
-    },
-    name: {
-      rich_text: [
-        {
-          text: {
-            content: name,
-          },
-        },
-      ],
-    },
-    date: {
-      date: {
-        start: new Date().toISOString(),
-      },
-    },
-  },
-  children: [
-    {
-      paragraph: {
-        rich_text: [
-          {
-            text: {
-              content,
-            },
-          },
-        ],
-      },
-    },
-    {
-      paragraph: {
-        rich_text: mentionPeople(),
-      },
-    },
-  ],
-});
-
-const createContact = async (
-  id: string,
-  email: string,
-  name: string,
-  content: string,
-) => {
-  try {
-    const response = await notion.pages.create(
-      createContactObject(id, email, name, content),
-    );
-
-    if (response.id && isFullPage(response)) {
-      return response.url;
-    }
-    throw {
-      body: {
-        message: "Failed to create notion page",
-      },
-    };
-  } catch (error) {
-    throw error;
-  }
-};
-
-const processContact = async (event: {
-  id: string;
-  email: string;
-  name: string;
-  message: string;
-}) => {
-  try {
-    const { id, email, name, message } = event;
-
-    if (!id || !email || !name || !message) {
-      throw {
-        body: {
-          message: "Missing id, email, name or message",
-        },
-      };
-    }
-
-    const url = await createContact(
-      `Message from ${name} (${id})`,
-      email,
-      name,
-      message,
-    );
-    await notifyContactCreated(name, email, url);
-  } catch (error) {
-    throw error;
-  }
-};
+const { NOTION_DATABASE_ID } = process.env;
 
 const allowRequest = async (request: Request & { ip?: string }) => {
   return { success: true, limit: 1, reset: Date.now() + 30000, remaining: 1 };
@@ -271,7 +33,7 @@ export async function OPTIONS() {
   });
 }
 
-export async function POST(request: Request) {
+export async function POST(request: NextRequest) {
   const corsHeaders = {
     "Access-Control-Allow-Origin": "*",
     "Access-Control-Allow-Credentials": "true",
@@ -330,6 +92,8 @@ export async function POST(request: Request) {
         email,
         name,
         message,
+        databaseID: NOTION_DATABASE_ID || "",
+        source: request.nextUrl.searchParams.get("source") || "Unknown",
       });
 
       return new Response(
